services/ai-engine-service/requirements.txt
# Generated from pyproject.toml - DO NOT EDIT MANUALLY
# Use scripts/generate-requirements.py to update

PyPDF2==3.0.1
aiohttp==3.9.0
argon2-cffi==23.1.0
asyncpg==0.29.0
chromadb==0.4.15
cryptography==45.0.6
fastapi==0.104.0
httpx==0.25.2
markdown==3.5.0
numpy==1.25.2
ollama==0.1.7
opentelemetry-api==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-sdk==1.21.0
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
pydantic-settings==2.0.0
pydantic[email]==2.5.0
python-docx==1.1.0
python-jose[cryptography]==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
redis==5.0.0
sqlalchemy==2.0.0
uvicorn[standard]==0.24.0